# packages/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Commodity Chain API
  version: 0.1.0
  description: |
    Contract-first API for the Commodity Chain regulated tokenized commodity exchange.
    Monetary amounts and commodity quantities are decimal integer strings expressed in
    the minor units identified by their accompanying currency or unit fields.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /v1
    description: Version 1 API base path
...
security:
  - OAuth2: []
paths:
...
  /oracle-events:
    post:
      tags:
        - Oracles
      summary: Accept a signed oracle event
      description: |
        Accepts an event signed by an approved oracle source. The X-Oracle-Signature
        header and body signature are both mandatory and must authenticate the same
        canonical event. The signed bytes are the UTF-8 encoding of the recursively
        key-sorted, whitespace-free JSON serialization of the request body after the
        top-level signature property has been removed. Array order is preserved.
        Ed25519 signatures use unpadded base64url, and X-Oracle-Signature must equal
        signature.value. Repeating an Idempotency-Key or the same sourceId/eventId
        returns the original HTTP status and result without another domain effect.
      operationId: acceptOracleEvent
      security:
        - OAuth2:
            - oracle:write
          OracleSourceSignature: []
      parameters:
        - $ref: '#/components/parameters/XCorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OracleEventEnvelope'
      responses:
        '202':
          description: |
            Oracle event applied, quarantined pending a missing nonce, or an original
            acceptance result replayed.
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/XCorrelationId'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OracleEventReceipt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      description: OAuth 2.0/OIDC bearer access token issued by the configured identity provider.
      flows:
        authorizationCode:
          authorizationUrl: https://identity.commodity-chain.local/realms/commodity-chain/protocol/openid-connect/auth
          tokenUrl: https://identity.commodity-chain.local/realms/commodity-chain/protocol/openid-connect/token
          scopes:
            oracle:write: Submit signed oracle events.
    OracleSourceSignature:
      type: apiKey
      in: header
      name: X-Oracle-Signature
      description: |
        Unpadded base64url Ed25519 signature over the canonical oracle event request.
        It must identify an approved source and exactly match signature.value in
        OracleEventEnvelope. Canonicalization recursively sorts object keys, preserves
        array order, emits no insignificant whitespace, and omits the top-level
        signature property before UTF-8 encoding.
  parameters:
    XCorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      description: Caller-supplied correlation identifier propagated to every response and event.
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Stable key for this command. Reusing the key with the same canonical request
        returns the original status and body without creating another effect. Reusing
        it with a different request returns IDEMPOTENCY_KEY_REUSED.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: ^[A-Za-z0-9._:-]+$
...
  headers:
    XCorrelationId:
      required: true
      description: Correlation identifier received with the request.
      schema:
        type: string
        format: uuid
    IdempotencyReplayed:
      required: true
      description: True when this is the stored response of an earlier request.
      schema:
        type: boolean
  responses:
    BadRequest:
      description: Request syntax, parameter, or schema validation failed.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication is required or the bearer token is invalid.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The authenticated subject is not permitted to perform this operation.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: The command conflicts with current state or an idempotency key payload.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: The request is syntactically valid but violates a domain invariant.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: An unexpected internal error occurred.
      headers:
        X-Correlation-Id:
          $ref: '#/components/headers/XCorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    NonNegativeIntegerString:
      type: string
      pattern: ^(0|[1-9][0-9]*)$
      description: Non-negative decimal integer string in the applicable minor units.
...
    UnitCode:
      type: string
      minLength: 1
      maxLength: 32
      pattern: ^[A-Z][A-Z0-9_]*$
      description: Controlled unit code identifying the minor units used for commodity quantities.
    Extensions:
      type: object
      description: Explicit namespace for backward-compatible, contract-governed extensions.
      additionalProperties: true
    InstrumentId:
      type: string
      format: uuid
      description: Internal instrument identifier.
    AssetId:
      type: string
      minLength: 1
      maxLength: 128
      pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]*$
      description: Internal or registry-qualified asset identifier.
...
    ExternalId:
      type: string
      minLength: 1
      maxLength: 128
      pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]*$
      description: Identifier assigned by an approved external system.
    Hash:
      type: string
      minLength: 16
      maxLength: 256
      description: Algorithm-qualified or otherwise unambiguous cryptographic digest.
    Signature:
      type: object
      additionalProperties: false
      required:
        - algorithm
        - keyId
        - value
      properties:
        algorithm:
          type: string
          minLength: 1
          maxLength: 64
        keyId:
          type: string
          minLength: 1
          maxLength: 128
        value:
          type: string
          minLength: 16
          maxLength: 8192
          description: Encoded detached or enveloped digital signature.
    OracleEd25519Signature:
      allOf:
        - $ref: '#/components/schemas/Signature'
        - type: object
          properties:
            algorithm:
              const: Ed25519
            value:
              type: string
              pattern: ^[A-Za-z0-9_-]+$
              description: Unpadded base64url Ed25519 signature bytes.
...
    OracleEventType:
      type: string
      enum:
        - COLLATERAL_RESERVED
        - RECEIPT_LOCKED
        - QUALITY_CONFIRMED
        - STOCK_UPDATED
        - GOODS_RELEASED
        - REVENUE_RECEIVED
    OracleProcessingStatus:
      type: string
      description: Current durable state of the oracle-event validation pipeline.
      enum:
        - RECEIVED
        - SCHEMA_VALIDATED
        - SIGNATURE_VALIDATED
        - POLICY_VALIDATED
        - APPLIED
        - DUPLICATE
        - STALE
        - QUARANTINED
        - REJECTED
    OracleEventEnvelope:
      type: object
      additionalProperties: false
      required:
        - eventId
        - schemaVersion
        - instrumentId
        - assetId
        - eventType
        - quantity
        - unit
        - observedAt
        - effectiveAt
        - sourceId
        - evidenceHash
        - nonce
        - signature
      properties:
        eventId:
          type: string
          format: uuid
        schemaVersion:
          type: string
          pattern: ^[1-9][0-9]*$
        instrumentId:
          $ref: '#/components/schemas/InstrumentId'
        assetId:
          $ref: '#/components/schemas/AssetId'
        eventType:
          $ref: '#/components/schemas/OracleEventType'
        quantity:
          allOf:
            - $ref: '#/components/schemas/NonNegativeIntegerString'
          description: Observed quantity as an integer string in the unit minor units.
        unit:
          $ref: '#/components/schemas/UnitCode'
        observedAt:
          type: string
          format: date-time
        effectiveAt:
          type: string
          format: date-time
        sourceId:
          $ref: '#/components/schemas/ExternalId'
        evidenceHash:
          $ref: '#/components/schemas/Hash'
        nonce:
          type: integer
          format: int64
          minimum: 0
          description: Monotonically increasing source nonce used with eventId for idempotency.
        signature:
          $ref: '#/components/schemas/OracleEd25519Signature'
        extensions:
          $ref: '#/components/schemas/Extensions'
    OracleEventReceipt:
      type: object
      additionalProperties: false
      required:
        - eventId
        - acceptedAt
        - status
        - replayed
      properties:
        eventId:
          type: string
          format: uuid
        acceptedAt:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/OracleProcessingStatus'
        replayed:
          type: boolean
...
    ErrorCode:
      type: string
      description: Stable, deterministic machine-readable error code.
      enum:
        - VALIDATION_ERROR
        - AUTHENTICATION_REQUIRED
        - PERMISSION_DENIED
        - RESOURCE_NOT_FOUND
        - CONFLICT
        - IDEMPOTENCY_KEY_REUSED
        - INVALID_STATUS
        - SUPPLY_EXCEEDS_COLLATERAL
        - SUPPLY_CAP_EXCEEDED
        - MINT_ACCOUNT_NOT_CONFIGURED
        - INSTRUMENT_STATE_INVALID
        - COLLATERAL_PROOF_INVALID
        - COLLATERAL_INSUFFICIENT
        - ORACLE_SIGNATURE_INVALID
        - ORACLE_SOURCE_UNKNOWN
        - ORACLE_SOURCE_KEY_REVOKED
        - ORACLE_NONCE_INVALID
        - ORACLE_NONCE_GAP
        - ORACLE_EVENT_STALE
        - ORDER_TYPE_NOT_AVAILABLE
        - ORDER_REJECTED
        - ORDER_NOT_CANCELLABLE
        - REDEMPTION_NOT_ALLOWED
        - INCOME_PROOF_INVALID
        - RATE_LIMITED
        - INTERNAL_ERROR
    ErrorDetail:
      type: object
      additionalProperties: false
      required:
        - reason
      properties:
        field:
          type: string
          minLength: 1
          maxLength: 256
        reason:
          type: string
          minLength: 1
          maxLength: 1000
        metadata:
          $ref: '#/components/schemas/Extensions'
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - correlationId
        - details
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          minLength: 1
          maxLength: 2000
        correlationId:
          type: string
          format: uuid
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'